+++ src/main/java/com/wenxin2/marioverse/entities/power_ups/MegaMushroomEntity.java
@@ -20,6 +20,7 @@ import software.bernie.geckolib.animation.AnimationController;
 import software.bernie.geckolib.animation.AnimationState;
 import software.bernie.geckolib.animation.PlayState;
 import software.bernie.geckolib.animation.RawAnimation;
+import software.bernie.geckolib.constant.DefaultAnimations;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
 public class MegaMushroomEntity extends MushroomEntity implements GeoEntity, PowerUpSource {
@@ -45,6 +46,7 @@ public class MegaMushroomEntity extends MushroomEntity implements GeoEntity, Pow
     @Override
     public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
         controllers.add(new AnimationController<>(this, "Walk", 0, this::walkAnimation));
+        controllers.add(DefaultAnimations.getSpawnController(this, state -> this, this.getSpawnDuration()));
     }
 
     @Override
@@ -69,5 +71,10 @@ public class MegaMushroomEntity extends MushroomEntity implements GeoEntity, Pow
             handler.applyMegaMushroomPowerUp(this.level(), rider, this);
         else if (entity instanceof LivingEntity livingEntity && entity instanceof AbilitiesHandler handler)
             handler.applyMegaMushroomPowerUp(this.level(), livingEntity, this);
+    }
+
+    @Override
+    protected int getSpawnDuration() {
+        return 25;
     }
 }