+++ src/main/java/com/wenxin2/marioverse/entities/power_ups/MiniMushroomEntity.java
@@ -18,6 +18,7 @@ import software.bernie.geckolib.animation.AnimationController;
 import software.bernie.geckolib.animation.AnimationState;
 import software.bernie.geckolib.animation.PlayState;
 import software.bernie.geckolib.animation.RawAnimation;
+import software.bernie.geckolib.constant.DefaultAnimations;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
 public class MiniMushroomEntity extends MushroomEntity implements GeoEntity, PowerUpSource {
@@ -36,6 +37,7 @@ public class MiniMushroomEntity extends MushroomEntity implements GeoEntity, Pow
     @Override
     public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
         controllers.add(new AnimationController<>(this, "Walk", 0, this::walkAnimation));
+        controllers.add(DefaultAnimations.getSpawnController(this, state -> this, this.getSpawnDuration()));
     }
 
     @Override