+++ src/main/java/com/wenxin2/marioverse/entities/power_ups/SuperMushroomEntity.java
@@ -19,6 +19,7 @@ import software.bernie.geckolib.animation.AnimationController;
 import software.bernie.geckolib.animation.AnimationState;
 import software.bernie.geckolib.animation.PlayState;
 import software.bernie.geckolib.animation.RawAnimation;
+import software.bernie.geckolib.constant.DefaultAnimations;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
 public class SuperMushroomEntity extends MushroomEntity implements GeoEntity, PowerUpSource {
@@ -37,6 +38,7 @@ public class SuperMushroomEntity extends MushroomEntity implements GeoEntity, Po
     @Override
     public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
         controllers.add(new AnimationController<>(this, "Walk", 0, this::walkAnimation));
+        controllers.add(DefaultAnimations.getSpawnController(this, state -> this, this.getSpawnDuration()));
     }
 
     @Override