+++ changelog.md
@@ -29,7 +29,7 @@
 - Improved Mini Goombas
   - The more goombas attached to the player, the slower you move
 - Added tooltips to warp pipes
-- Added spawn animations to power ups
+- Added new spawn animations to power ups
 - Added flame particles to the Fire Flower
 - Added ice particles to the Ice Flower
 - Remodeled the fire & ice flower power ups